Basics_flask/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-wtf==1.1.1
email_validator==2.0.0
werkzeug==2.2.3
pandas==1.5.3
python-dotenv==1.0.0